+++ src/manage.c
@@ -46,6 +46,7 @@
  */
 #define _GNU_SOURCE
 
+#include "gmp_base.h"
 #include "manage.h"
 #include "manage_acl.h"
 #include "manage_configs.h"
@@ -4614,15 +4615,19 @@ run_osp_task (task_t task, int from, char **report_id)
  * @brief Perform a CVE "scan" on a host.
  *
  * @param[in]  task      Task.
+ * @param[in]  report    The report to add the host, results and details to.
  * @param[in]  gvm_host  Host.
  *
  * @return 0 success, 1 failed to get nthlast report for a host.
  */
 static int
-cve_scan_host (task_t task, gvm_host_t *gvm_host)
+cve_scan_host (task_t task, report_t report, gvm_host_t *gvm_host)
 {
   report_host_t report_host;
   gchar *ip, *host;
+
+  assert (task);
+  assert (report);
 
   host = gvm_host_value_str (gvm_host);
 
@@ -4632,7 +4637,7 @@ cve_scan_host (task_t task, gvm_host_t *gvm_host)
 
   g_debug ("%s: ip: %s", __func__, ip);
 
-  /* Get the last report that applies to the host. */
+  /* Get the last report host that applies to the host IP address. */
 
   if (host_nthlast_report_host (ip, &report_host, 1))
     {
@@ -4664,11 +4669,13 @@ cve_scan_host (task_t task, gvm_host_t *gvm_host)
             {
               const char *app, *cve;
               double severity;
-              gchar *desc, *location;
+              gchar *desc;
+              iterator_t locations_iter;
+              GString *locations;
               result_t result;
 
-              if (global_current_report && (prognosis_report_host == 0))
-                prognosis_report_host = manage_report_host_add (global_current_report,
+              if (prognosis_report_host == 0)
+                prognosis_report_host = manage_report_host_add (report,
                                                                 ip,
                                                                 start_time,
                                                                 0);
@@ -4677,7 +4684,34 @@ cve_scan_host (task_t task, gvm_host_t *gvm_host)
 
               app = prognosis_iterator_cpe (&prognosis);
               cve = prognosis_iterator_cve (&prognosis);
-              location = app_location (report_host, app);
+              locations = g_string_new("");
+
+              insert_report_host_detail (global_current_report, ip, "cve", cve,
+                                         "CVE Scanner", "App", app);
+
+              init_app_locations_iterator (&locations_iter, report_host, app);
+
+              while (next (&locations_iter))
+                {
+                  const char *location;
+                  location = app_locations_iterator_location (&locations_iter);
+
+                  if (locations->len)
+                    g_string_append (locations, ", ");
+                  g_string_append (locations, location);
+
+                  insert_report_host_detail (report, ip, "cve", cve,
+                                             "CVE Scanner", app, location);
+
+                  insert_report_host_detail (report, ip, "cve", cve,
+                                             "CVE Scanner", "detected_at",
+                                             location);
+
+                  insert_report_host_detail (report, ip, "cve", cve,
+                                             "CVE Scanner", "detected_by",
+                                             /* Detected by itself. */
+                                             cve);
+                }
 
               desc = g_strdup_printf ("The host carries the product: %s\n"
                                       "It is vulnerable according to: %s.\n"
@@ -4686,11 +4720,11 @@ cve_scan_host (task_t task, gvm_host_t *gvm_host)
                                       "%s",
                                       app,
                                       cve,
-                                      location
+                                      locations->len
                                        ? "The product was found at: "
                                        : "",
-                                      location ? location : "",
-                                      location ? ".\n" : "",
+                                      locations->len ? locations->str : "",
+                                      locations->len ? ".\n" : "",
                                       prognosis_iterator_description
                                        (&prognosis));
 
@@ -4699,28 +4733,10 @@ cve_scan_host (task_t task, gvm_host_t *gvm_host)
 
               result = make_cve_result (task, ip, cve, severity, desc);
               g_free (desc);
-              if (global_current_report)
-                {
-                  report_add_result (global_current_report, result);
 
-                  insert_report_host_detail (global_current_report, ip, "cve", cve,
-                                             "CVE Scanner", "App", app);
+              report_add_result (report, result);
 
-                  if (location)
-                    {
-                      insert_report_host_detail (global_current_report, ip, "cve", cve,
-                                                 "CVE Scanner", app, location);
-
-                      insert_report_host_detail (global_current_report, ip, "cve", cve,
-                                                 "CVE Scanner", "detected_at",
-                                                 location);
-                      insert_report_host_detail (global_current_report, ip, "cve", cve,
-                                                 "CVE Scanner", "detected_by",
-                                                 /* Detected by itself. */
-                                                 cve);
-                    }
-                }
-              g_free (location);
+              g_string_free (locations, TRUE);
             }
           cleanup_iterator (&prognosis);
 
@@ -4729,7 +4745,7 @@ cve_scan_host (task_t task, gvm_host_t *gvm_host)
               /* Complete the report_host. */
 
               report_host_set_end_time (prognosis_report_host, time (NULL));
-              insert_report_host_detail (global_current_report, ip, "cve", "",
+              insert_report_host_detail (report, ip, "cve", "",
                                          "CVE Scanner", "CVE Scan", "1");
             }
         }
@@ -4826,7 +4842,7 @@ fork_cve_scan_handler (task_t task, target_t target)
   gvm_hosts = gvm_hosts_new (hosts);
   free (hosts);
   while ((gvm_host = gvm_hosts_next (gvm_hosts)))
-    if (cve_scan_host (task, gvm_host))
+    if (cve_scan_host (task, global_current_report, gvm_host))
       {
         set_task_interrupted (task,
                               "Failed to get nthlast report."
@@ -8036,21 +8052,21 @@ get_nvt_xml (iterator_t *nvts, int details, int pref_count,
           nvt_iterator_solution_type (nvts) ||
           nvt_iterator_solution_method (nvts))
         {
-          g_string_append_printf (buffer, "<solution");
+          buffer_xml_append_printf (buffer, "<solution");
 
           if (nvt_iterator_solution_type (nvts))
-            g_string_append_printf (buffer, " type='%s'",
+            buffer_xml_append_printf (buffer, " type='%s'",
               nvt_iterator_solution_type (nvts));
 
           if (nvt_iterator_solution_method (nvts))
-            g_string_append_printf (buffer, " method='%s'",
+            buffer_xml_append_printf (buffer, " method='%s'",
               nvt_iterator_solution_method (nvts));
 
           if (nvt_iterator_solution (nvts))
-            g_string_append_printf (buffer, ">%s</solution>",
+            buffer_xml_append_printf (buffer, ">%s</solution>",
               nvt_iterator_solution (nvts));
           else
-            g_string_append_printf (buffer, "/>");
+            buffer_xml_append_printf (buffer, "/>");
         }